Workaround for Ruby UTC offset bug
Some Ruby versions have a bug when creating a Time object backed by a
timezone object where they create a fractional-second UTC offset.

For the added test, without this workaround, on Ruby 3.3.0:

      1) Failure:
    TimeExtCalculationsTest#test_change_preserves_fractional_seconds_on_zoned_time [./test/core_ext/time_ext_test.rb:528]:
    --- expected
    +++ actual
    @@ -1 +1,3 @@
    -"2005-01-30 00:00:00.99 -0500"
    +# encoding: US-ASCII
    +#    valid: true
    +"2005-01-30 00:00:00.99 -045959"

This is fixed in Ruby 3.3.1 and 3.2.4. We can remove the workaround when
we expect users to be on those versions or newer.